fix: PR #116 のレビューコメントを反映
- ConfigResponseのvalue型をstring | number | boolean | Record<string, unknown>に修正
- エラーコードにINVALID_KEYを追加
- 重複したconvertConfigValue関数をutils.tsに抽出してDRY原則に準拠
- switchステートメントの変数スコープ問題を修正
- 安全でない型アサーションにバリデーションを追加

AI Agents prompts に基づく自動修正
